backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pandas==2.1.3
python-dotenv==1.0.0
cryptography==41.0.7
python-multipart==0.0.6 